End of Day 2. Implemented Main Component. Working on updating state with api on main card with preferences. 5/7/22
